+++ changelog.md
@@ -6,6 +6,7 @@
 * `CHG` supports `~` in command line
 * `CHG` completion: improve workspace words
 * `CHG` completion: show words in string
+* `CHG` completion: split `for .. in` to `for .. ipairs` and `for ..pairs`
 * `FIX` [#339](https://github.com/sumneko/lua-language-server/issues/339)
 
 ## 1.9.0